RISC-V: Always register vector built-in function at LTO pipeline [PR110812]
This patch ensures that vector built-in functions are registered
during the LTO pipeline, addressing an issue where these functions were
not being recognized in certain compilation scenarios.

And we already check the extension requirement at expand time, so
it's safe to alwasy register them.

The price of this change is it will increase the bootstrap time for LTO.